got json and bots on board
Created a format to save shows in a file.
Started playback by putting robots on board.
Need to make a function to decode the spline into xy positions for the simulator to use.
Also need to create a start/stop button and playback bar.
Alternatively, we could use drive/rotate packets to store all the different commands the robot will need to get to their destination. It would be simpler but would not be able to encode a curve. idk what's better